LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssTransformValue.tentative-expected.txt


PASS Constructing a CSSTransformValue with no components throws TypeError
PASS CSSTransformValue can be constructed with multiple transforms
PASS CSSTransformValue.set correctly sets the CSSTransformComponent at the given index
PASS Setting a component in CSSTransformValue correctly appends the CSSTransformComponent if index specified is greater than length
PASS CSSTransformValue.is2D is false when given mix of 2D and 3D transforms
PASS CSSTransformValue.is2D is true when given only 2D transforms
PASS CSSTransformValue.is2D is readonly
PASS Can iterate through CSSTransformValue components
